+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Added support for YAML-based validation files in DevContext (https://github.com/authzed/spicedb/pull/3024)
+
 ### Changed
 - Added classic histogram buckets to the Perf Insights query latency metric for better slow-query visibility (https://github.com/authzed/spicedb/pull/3027)
 - Removed MySQL metrics prefixed with `go_sql_stats_connections_*` in favor of those prefixed with `go_sql_*` (https://github.com/authzed/spicedb/pull/2980)